fix: added softstart for application
fix: added close window functionality to overview
