+++ crates/windows-lib/src/overview/mod.rs
@@ -1,10 +1,13 @@
 mod close;
+mod close_item;
 mod create;
 mod open;
+mod render;
 mod stop;
 mod update;
 
 pub use close::*;
+pub use close_item::*;
 pub use create::*;
 pub use open::*;
 pub use stop::*;
